pyproject.toml
[tool.poetry]
name = "ur-mediathek-downloader"
version = "0.1.0"
description = "Tools to download stuff from the UR mediathek"
authors = ["RDSK"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
selenium-wire = "^2.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
